requirements.txt
fastapi
uvicorn[standard]
python-docx
lxml
openai
python-multipart
aiofiles
